added correct implementation of tree

+++ algo-testing/src/com/keybase/avminin/Utils.java
@@ -1,19 +1,19 @@
 package com.keybase.avminin;
 
 public class Utils {
-    public static boolean gt(int a, int b) {
+    public static boolean gt(double a, double b) {
         return a > b;
     }
 
-    public static boolean lt(int a, int b) {
+    public static boolean lt(double a, double b) {
         return a < b;
     }
 
-    public static int mean(int a, int b) {
+    public static double mean(double a, double b) {
         return (a + b) / 2;
     }
 
-    public static int signum(int a, int b) {
+    public static int signum(double a, double b) {
         if (a == b) return 0;
         return a < b ? -1 : 1;
     }
