+++ algo-testing/src/com/keybase/avminin/heaps/HeapADT.java
@@ -1,8 +1,8 @@
 package com.keybase.avminin.heaps;
 
 public interface HeapADT {
-    boolean Insert(int e);
-    int  GetTop();
-    int  ExtractTop();
+    boolean Insert(double e);
+    double  GetTop();
+    double  ExtractTop();
     int  GetCount();
 }
