+++ algo-testing/src/com/keybase/avminin/heaps/MaximumHeap.java
@@ -5,6 +5,6 @@ import com.keybase.avminin.Utils;
 
 public class MaximumHeap extends MainHeap {
     public MaximumHeap() {
-        super(new int[MAX_HEAP_SIZE], Utils::gt);
+        super(new double[MAX_HEAP_SIZE], Utils::gt);
     }
 }
