+++ algo-testing/src/com/keybase/avminin/heaps/MinimumHeap.java
@@ -4,6 +4,6 @@ import com.keybase.avminin.Utils;
 
 public class MinimumHeap extends MainHeap {
     public MinimumHeap() {
-        super(new int[MAX_HEAP_SIZE], Utils::lt);
+        super(new double[MAX_HEAP_SIZE], Utils::lt);
     }
 }
